classe configura fornecimento

+++ Projeto_site/src/com/jor/site/controle/ProdutoFornecidoControler.java
@@ -1,11 +1,13 @@
 package com.jor.site.controle;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.hibernate.Criteria;
 import org.hibernate.Query;
 import org.hibernate.Session;
 
+import com.jor.site.entidade.Cliente;
 import com.jor.site.entidade.ProdutoFornecido;
 import com.jor.site.util.Alerta;
 import com.jor.site.util.HibernateUtil;
@@ -32,10 +34,15 @@ public class ProdutoFornecidoControler {
 	public void deletar(ProdutoFornecido ProdutoFornecido)
 	{
 		session = HibernateUtil.getSessionFactory().openSession();
-		try{			
-			session.beginTransaction();
-			session.delete(ProdutoFornecido);
-			session.getTransaction().commit();
+		try{
+			
+			if (session.getTransaction() != null && session.getTransaction().isActive()) {
+			   session.getTransaction().commit();			   
+			} else{
+				session.beginTransaction();
+				session.delete(ProdutoFornecido);			
+				session.getTransaction().commit();
+			}
 		}
 		catch(ExceptionInInitializerError er)
 		{			
@@ -45,6 +52,7 @@ public class ProdutoFornecidoControler {
 		{
 			Alerta.info("ProdutoFornecido "+ProdutoFornecido.getNome()+" deletado com sucesso");
 			session.close();
+			System.out.println("sesao foi fechada para deletar");
 		}
 	}
 	
@@ -53,20 +61,26 @@ public class ProdutoFornecidoControler {
 	public List buscaProdutoFornecidos(String data,int id) {
 		session = HibernateUtil.getSessionFactory().openSession();
 		try {
-	    	
+	    	System.out.println("sesao foi aberta para pesquisar");
 	    	session.beginTransaction();
 	        Query q = session.createQuery ("from ProdutoFornecido where dataCadastro = '"+data+"' and "
 	        		                      + "id_parceria ='"+id+"'");
 	        	       
-	        return q.list();
-	        
+	        List list = q.list();
+	        session.close();
+	        System.out.println("sesao foi fechada para pesquisar");
+	        if(!list.isEmpty())
+	         return list;	
+	        else
+	         return new ArrayList();
 	    } catch (Exception e) {
 	         System.out.println("erro ao pesquisar p "+e.getMessage());
 	    }
-		finally{
-			session.close();
-		}
+		
+			
+		
 		return null;
-	}	
+	}
+	
 	
 }
